Implement fullscreen mode. Yeah, this takes the new floating toolbar in action :)
The new fullscreen mode is completely different implemented than the old one was. Now we use Qt / KDE fullscreen mode; but not change the resolution of the x-server anymore, which was IMHO very ugly.

Also add some initial actions for remote views. Nice extra: Take a screenshot of the remote desktop and send it directly to the clipboard.

svn path=/branches/work/soc-krdc/krdc/; revision=673068