Backport: SVN commit 854023 by gpothier:
Fix concurrency issues. In particular this should fix crashes when a
connection is closed either from krdc or from the server side.
Bug reporters, please check if you can still reproduce the crashes with this
patch (this is for trunk, backporting to 4.1 in a few minutes).

Details:
1- Disconnect signals from the client thread to the vncview when the
latter is deleted. This Fixes a 100% repeatable crash that occurs when
I close a vnc tab that is connected to an ubuntu machine running the
vino vnc server (aka gnome desktop sharing)
2- Remove the buf global variable. I wasn't able to get a
reproducible crash due only to this variable because of the above
item, but getting rid of that variable is the sanest thing to do I
think.

CCMAIL:gpothier@gmail.com

svn path=/branches/KDE/4.1/kdenetwork/krdc/; revision=854129