RDP: remember the resolution selection
Currently, we use the saved width and height to calculate
the resolution selection, but it's not always working.
When user choose "Current KRDC Size", it'll become
"Custom Resolution (...)" with both width and height are set to 0
when the next time we open the host configuration dialog.

It's very annoying when we prefer to use "Current KRDC Size". So,
save the resolution selection is a better way to tackle with this issue.
And we don't need the hack that maps width/height to resolution selection.

BUG: 337549

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>