Fix handling command line arguments
QUrl::fromLocalFile is clearly wrong.
It gives us file://rdp:// or similar garbage.

We cannot use QUrl::fromUserInput() either, since this gives
us an http:// or an ftp:// URL if the scheme is empty,
depending on the hostname. Due to the plugin architecture
we cannot assume that krdc can't handle those protocols.

Assume that we're dealing with just a hostname instead,
and further assume that the protocol is vnc://
(default since KDE3).

REVIEW: 128060