Make default protocol configurable
Add a new config option for the default protocol.
Default protocol: protocol to use when passing a hostname
via command line without a scheme.
~~
$ krdc my.host
~~
expands to vnc://my.host since the dawn of time. With the
default protocol set to "rdp" it expands to rdp://my.host

The default protocol defaults to vnc to keep the current
behavior.

REVIEW: 128496