Signal the main window about vnc disconnection
Summary:
When the client is quitted and have stopped trying to reconnect.

Previously, a failed connection will leave the main window in a "connected" state
and require manually pressing the "Disconnect" button before a new connection can be made.

Use of the `disconnected` signal copied from the RDP backend, which handles this correctly.

Test Plan: Tested behavior on 20.04 branch. Tested compilation also on master.

Reviewers: davidedmundson, uwolfer, ahmadsamir, murrant, cfeck

Reviewed By: uwolfer

Differential Revision: https://phabricator.kde.org/D29822