Add highdpi support for RDP and VNC
Summary:
In highdpi mode coordinates are specified in device independent pixels,
and Qt auto-scales everything by devicePixelRatio, i.e. the ratio between
the device independent and device pixel coordinate systems.

Thus for RDP we should pass the size in device pixels to xfreerdp.

For VNC we should avoid auto-scaling, which can be done by setting
devicePixelRatio of the paint device for the painted image, and
the size of the image should be converted to device independent
pixels where it's needed.

Reviewers: #kde_applications, davidedmundson, uwolfer, ahmadsamir, murrant, cfeck, ngraham

Reviewed By: murrant

Subscribers: yuyichao, ngraham, alexeymin, aacid

Differential Revision: https://phabricator.kde.org/D20790