Fix low quality connections
If you try to create a VNC remote connection with the "low quality" setting,
only a blue or black screen is shown. The reason is libvnc which does not
support tight encoding and 8 bits per pixel at the same time. Only 8 bpp
without tight encoding or tight encoding with 16 or 32 bpp is possible.

This patch removes tight encoding from low quality connection configuration.