revise VNC/RDP option selection
the RDP side of things was actually confusingly wrong as it was tagged
RUNTIME but was required at buildtime. this was because we have the
buildtime stuff doubling as feature selectors. this is a bit meh and
indeed a similar case was discussed a while ago in #kde-devel with the
conclusion mostly being that optional deps are meh and opting out of
stuff is probably much more expressive.

so. this commit make the user opt out of features rather than enabling
them when dependencies are available.

cmake options WITH_VNC and WITH_RDP now control the relevant feature
sets. they are both enabled by default. when disabled the relevant deps
aren't being looked for anymore, still enabling the user to opt out of a
given feature set.
this then also frees up freerdp's package to be proper RUNTIME because
the feature enabling is controlled by the option, not the presence.

both options also express themselves in the feature summary to further
aid expressiveness and discoverability of the options.

LibVNCServer is now a REQUIRED dep of the VNC feature. libssh continues
to be OPTIONAL

freerdp also has a proper program finder added to get rid of the awkward
hack in the cmakelists.txt that did largely the same anyway