Replace functions deprecated in Qt 5.15
This patch replaces:

* vsprintf with vasprintf.
* functions like `QApplication::desktop()->screenGeometry()`

I've done some tests, and it seems to work properly.

As mentioned at https://invent.kde.org/network/krdc/-/issues/4#note_433301 there are some use cases that need attention, but it shouldn't be done in the scope of this MR as the problem exists as well with the deprecated functions.