Fix krdc-vnc-qtonly
krdc-vnc-qtonly is a very simple VNC client that's part of KRDC. In its current state, it has several problems:

1. It uses qmake.
2. It doesn't build at all.
3. It's very slightly bit-rotten.
4. It requires copying files manually from other parts of KRDC in order to build it.

This MR fixes all four. The project is ported to CMake, made able to build, the one bit-rotten problem in the used KRDC code is fixed, and the new CMake code automatically includes the needed files without requiring manual copying.

Note that I am very to using CMake and mostly copied from other CMakeLists.txt files in KRDC to do the port. The project builds and runs properly, but it may need some changes.