Replace RDP plugin with a new plugin that uses FreeRDP as a library
Currently the RDP plugin uses xfreerdp as separate executable and embeds
it using some X trickery. This does not work on wayland and has no
chance to work on wayland. Instead, we can use FreeRDP directly as a
library and use it to render to a widget directly. In addition, while
we're at it, we can modernise some parts of the RDP support like adding
support for the graphics pipeline extension.