Add draft README with detailed documentation
This merge request adds an initial draft of the README file for the KRDC project. The README includes installation instructions, usage, contributing guidelines, and additional links to relevant documentation for users and developers. This draft aims to improve the accessibility of project information and provide clear guidance for potential contributors.

The README is based on the existing documentation and includes links to:
- [Official KRDC website](https://apps.kde.org/krdc/)
- [KDE Documentation](https://docs.kde.org/stable5/en/krdc/krdc/index.html)