Fix scaling button not working out of the box
Note that this only seems to affect VNC connections, because it has its
own scaling setting.

There has been this bug for years now, where if you press the "Scale"
button in the toolbar that the image disappears. What is actually
happening is that there is a "scale factor" setting, which ranges from
100 to 200. (This is translated internally to 1x and 2x, respectively.)
The *default value* for this setting however was 0, which meant that it
scaled to the view to basically nothing. This makes absolutely no sense
as a default value, so it's changed to 100. As a workaround, you can
also change the scale factor using the slider, it's shown by default in
the fullscreen toolbar.

However, this fix doesn't change it for many users - who have collected
numerous VNC connections like me - and the scale setting is still 0.
Thankfully because this value is nonsense and can't be set in the UI, we
can detect it and remove it. I have tested this against a 24.12 KRDC and
can confirm it fixes scaling for connections made in that version.

Additionally, I have amended the doucmentation for the scaleFactor
setting so it's absolutely clear to everybody what the valid range for
it is.

BUG: 485094
VERSION-FIXED-IN: 25.04


(cherry picked from commit ae2f8d4550d78ef975369cd59696099a86efbe03)

Co-authored-by: Joshua Goins <joshua.goins@kdab.com>